requirements.txt
customtkinter>=5.2.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pywin32>=306
Pillow>=10.0.0
pyinstaller>=6.0.0
